mysite/requirements.txt
beautifulsoup4==4.6.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-extensions==2.1.4
django-modelcluster==4.3
django-taggit==0.23.0
django-tempus-dominus==5.1.2.1
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
draftjs-exporter==2.1.5
graphviz==0.10.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
mysqlclient==1.4.1
networkx==2.2
oauthlib==2.1.0
Pillow==5.4.1
pydotplus==2.0.2
pyparsing==2.3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.1.0
six==1.12.0
Unidecode==1.0.23
urllib3==1.24.1
wagtail==2.4
webencodings==0.5.1
Willow==1.1
